IOS48 : update Cheq API, split DateHelperUtil into TestUtil , fixed error with VDotManager

+++ Cheq/Cheq/Service/CheqAPI/v1/Models/UserBudgetUpdate.swift
@@ -13,9 +13,9 @@ public struct UserBudgetUpdate: Codable {
 
     public var budgetId: Int?
     public var hide: Bool?
-    public var newEstimatedBudget: Double?
+    public var newEstimatedBudget: Int?
 
-    public init(budgetId: Int?, hide: Bool?, newEstimatedBudget: Double?) {
+    public init(budgetId: Int?, hide: Bool?, newEstimatedBudget: Int?) {
         self.budgetId = budgetId
         self.hide = hide
         self.newEstimatedBudget = newEstimatedBudget
